requirements.txt
robotframework~=3.2.2
robotframework-sshlibrary
paramiko~=2.7.2
matplotlib
pandas
PyYAML~=5.4.1
setuptools~=54.1.2